feat: Renderables are entity components
